# Speed up fuzzy search