--- a/ghcide/ghcide.cabal
+++ b/ghcide/ghcide.cabal
@@ -50,7 +50,6 @@ library
         dlist,
         exceptions,
         extra >= 1.7.4,
-        fuzzy,
         filepath,
         fingertree,
         focus,