--- a/ghcide/src/Text/Fuzzy/Parallel.hs
+++ b/ghcide/src/Text/Fuzzy/Parallel.hs
@@ -3,6 +3,7 @@
 module Text.Fuzzy.Parallel
 (   filter,
     simpleFilter,
+    match,
     Scored(..)
 ) where