My intuition is that this approach should be faster, but I do not have data to back such claim. @pepeiborra any chance to test it with a huge list of completions?

<a href="https://gitpod.io/#https://github.com/haskell/haskell-language-server/pull/2639"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>