@pepeiborra 2022-01-25
Thanks @Bodigrim (and sorry to have dragged you into another project :) )

I have a benchmark in the Sigma codebase for this completions fuzzy search, but I will need a few days to find the time. Probably over the weekend.