@Bodigrim 2022-01-25
I've probably broken something, but it's still surprising that 9.0.1 Ubuntu and 9.2.1 Mac pass, while 9.2.1 Ubuntu fails.