@Bodigrim 2022-01-27
https://github.com/haskell/haskell-language-server/runs/4959105656?check_suite_focus=true#step:7:9685 says
> Exception: Prelude.head: empty list
So helpful! Cannot wait for GHC 9.4 with stack traces for `head`.