@pepeiborra 2022-01-29
Benchmarked using the completions experiment of the ghcide bench suite on the Sigma codebase. The results show that the new code allocates much less and is up to 2X as fast. Nice job!

### BEFORE
```
name        | success | samples | startup | setup | userTime | delayedTime | totalTime
----------- | ------- | ------- | ------- | ----- | -------- | ----------- | ---------
completions | True    | 100     | 1m25s   | 0.00s | 1m08s    | 0.05s       | 1m08s    
/data/users/pepeiborra/fbsource/fbcode/buck-out/opt/gen/sigma/ide/sigma-ide --lsp --test --cwd /home/pepeiborra/si_sigma +RTS -A32M -I0 -s/tmp/sigma-ide-benchmark-large.gcStats 
 641,948,658,384 bytes allocated in the heap
  74,949,086,032 bytes copied during GC
  10,809,313,048 bytes maximum residency (14 sample(s))
     645,310,696 bytes maximum slop
           10308 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       979 colls,   978 par   1089.292s  27.336s     0.0279s    0.5449s
  Gen  1        14 colls,    12 par   351.536s  17.144s     1.2245s    7.6997s

  Parallel GC work balance: 84.43% (serial 0%, perfect 100%)

  TASKS: 215 (6 bound, 209 peak workers (209 total), using -N40)

  SPARKS: 46800 (46573 converted, 0 overflowed, 0 dud, 0 GC'd, 227 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  718.015s  (117.626s elapsed)
  GC      time  1440.828s  ( 44.479s elapsed)
  EXIT    time    0.046s  (  0.005s elapsed)
  Total   time  2158.890s  (162.111s elapsed)

  Alloc rate    894,060,716 bytes per MUT second

  Productivity  33.3% of total user, 72.6% of total elapsed
```

### AFTER
```
name        | success | samples | startup | setup | userTime | delayedTime | totalTime
----------- | ------- | ------- | ------- | ----- | -------- | ----------- | ---------
completions | True    | 100     | 1m20s   | 0.00s | 43.60s   | 0.05s       | 43.67s   
/data/users/pepeiborra/fbsource/fbcode/buck-out/opt/gen/sigma/ide/sigma-ide --lsp --test --cwd /home/pepeiborra/si_sigma +RTS -A32M -I0 -s/tmp/sigma-ide-benchmark-large.gcStats 
 351,299,429,720 bytes allocated in the heap
  47,676,026,872 bytes copied during GC
  10,489,733,720 bytes maximum residency (13 sample(s))
     642,772,392 bytes maximum slop
           10003 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       757 colls,   756 par   487.501s  12.327s     0.0163s    0.5824s
  Gen  1        13 colls,    11 par   294.023s  15.454s     1.1888s    7.6197s

  Parallel GC work balance: 84.19% (serial 0%, perfect 100%)

  TASKS: 218 (6 bound, 212 peak workers (212 total), using -N40)

  SPARKS: 46800 (46670 converted, 0 overflowed, 0 dud, 0 GC'd, 130 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time  404.382s  (104.802s elapsed)
  GC      time  781.524s  ( 27.781s elapsed)
  EXIT    time    0.073s  (  0.008s elapsed)
  Total   time  1185.980s  (132.593s elapsed)

  Alloc rate    868,731,964 bytes per MUT second

  Productivity  34.1% of total user, 79.0% of total elapsed
```