@Bodigrim 2022-02-04
@pepeiborra just a gentle reminder about my questions above.